Simplify
Since the GITHUB_TOKEN for the action doesn't have the correct scope it
can't access team membership. We don't want to hard code the team
members, so we'll have to figure out a non-hacky way to filter these.

We may need to set up a properly scoped PAT, but that's a bit of a pain
given the security posture.