Automatically comment on CODEOWNERS Pull Requests (#4)
This PR adds a new organization workflow utilizing the github-script
action that will add comments to PRs on the following events:

1. A review request (pull_request.review_requested) from a given CODEOWNERS
   github team (release-engineering-reviewers).
2. A user adds the "ready for RE review" label (pull_request.review_requested)
   to a pull request.

We also tried a graphql query to combine multiple multiple REST calls to filter
reviewers to handle auto-assignment of PRs, but since the GITHUB_TOKEN for the
action doesn't have the correct scope it can't access team membership. We don't
want to hard code the team members, so we'll have to figure out a non-hacky way
to filter these.

We may need to set up a properly scoped PAT, but that's a bit of a pain
given the security posture.